+++ src/duckylib/source/cmath.c
@@ -1,6 +1,0 @@
-#include "../utils/cmath.h"
-
-float cmath_to_radians(float value)
-{
-    return value * M_PI / 180;
-}