+++ src/main.c
@@ -4,22 +4,22 @@
 
 int main(int argc, char *argv[])
 {
-
-    printf("hi mum!!\n");
-
     DL_Window window;
-    dl_startup(&window, true, true);
+    dl_startup(&window, "DuckyLib", true, true);
 
     DL_UIEntity ui = dl_ui_entity_new(NULL, BM_NEAREST, DL_Aspect_1920x1080, NULL, NULL);
-    ui.scale.x = 0.5f;
-    ui.scale.y = 0.5f;
+    // ui.scale.y = 0.5f;
     dl_renderer_unbind_all(&ui.renderer);
+
+    DL_UIEntity ui2 = dl_ui_entity_new(DL_DEFAULT_TEXTURE, BM_NEAREST, DL_Aspect_1920x1080, NULL, NULL);
+    ui2.scale.x = 10.0f;
+    ui2.scale.y = 10.0f;
+    dl_renderer_unbind_all(&ui2.renderer);
 
     printf("%f\n", ui.aspect_ratio);
 
-    DL_Bind msg = {DL_LMB};
-
-    printf("%f, %f, %f\n", ui.position.x, ui.position.y, ui.position.z);
+    DL_Bind lmb = {DL_LMB};
+    DL_Bind spc = {DL_SPACE};
 
     while (window.running)
     {
@@ -33,19 +33,36 @@ int main(int argc, char *argv[])
 
         dl_camera_update();
         dl_ui_entity_update(&ui);
+        dl_ui_entity_update(&ui2);
+
+        // if(dl_ui_entity_is_mouse_over(&ui, &window))
+        // {
+
+        // }
+
+        if(dl_input_get_key_down(&spc, true))
+        {
+            dl_log_new("yo", DL_MSG);
+        }
+
+        if (dl_input_get_key_down(&lmb, false))
+        {
+            // float x = dl_input_get_mouse_position().x;
+            // x -= window.viewport_x;
+            // x -= window.viewport_w / 2;
+            // x /= window.viewport_w / 2;
+            // float y = dl_input_get_mouse_position().y;
+            // y -= window.viewport_y;
+            // y -= window.viewport_h / 2;
+            // y /= window.viewport_h / 2;
+            // ui.position.x = x;
+            // ui.position.y = -y;
+
+            ui.position = dl_input_get_mouse_position(&window);
+        }
 
         dl_window_swap_buffer(&window);
         dl_frame_end();
-
-        if(dl_ui_entity_is_mouse_over(&ui, &window))
-        {
-
-        }
-
-        if(dl_input_get_key_down(&msg, true))
-        {
-            printf("yo yo yo\n");
-        }
     }
     dl_window_destroy(&window);
     dl_camera_destroy();